Commit on Exercise 1.1, update .gitignore file.
It is the first formal commit on my SICP learning project, included are the following 3 files:

   Exercise_1.1.tm
   Exercise_1.1.pdf
   .gitignore

the first one is a TeXmacs source file for the problem and solution. .gitignore has been updated,
so I push it up as well as the new files.

From this commit on, I'll keep pushing my exercises in SICP to the repository regularly.

Thank you!

											Lawrence R. Amlord
											  Mar. 19, 2013
											  Xi'an, China