Add my answer to Exercise 1.19.
Included are the following 3 files:

	Exercise_1.19.pdf
	Exercise_1.19.scm
	Exercise_1.19.tm

By a first glance, one might superficially conclude that this exercise is
easy--with only 2 slots to fill in, however, this really isn't the case.
This exercise requires the reader a careful observation, a maturity of
mathematical deduction and the ability to deliberate it on different levels
of abstraction. But don't be intemidated by its difficulty, you ability of
abstract thinking will be improved as long as you are willing to involve in
interacting with it.

All in all, it is worthwhile devoting a great efforts to tackle this chal-
lenge.

Thank you!
							Lawrence R. Amlord
							   Apr 28, 2013
							   Xi'an, China