Add my answer to Exercise 1.31.
Included are the following 4 files:

	Exercise_1.31.pdf
	Exercise_1.31.scm
	Exercise_1.31.tm
	Test_for_Exercise_1.31.scm

Note that here the implementation of procedure `pi-quotient' is not pretty
nice. For it is even unable to compute the approximation of Pi when scale of
the last factor of numerator becomes greater than 175. I believe there should
exist ways to improve this implementation.

If you have any good idea or find any error in my solution. Please contact me
through email: informmarry@gmail.com

Thank you!
							Lawrence R. Amlord
							   May 9th, 2013
							   Xi'an, China