add all code of the book the authors provided.
All code in the text are packed into a folder named 'allcode'.
Some of them simply collect code arise in each chapter:

	ch1.scm
	ch2.scm
	ch3.scm
	ch4.scm
	ch5.scm

They are not expected to run as a whole.

The rest part of the package presents supporting
module and implements interpreters/compilers
described in the text in order to facilitate the
reader to explore the examples and exercises.

	*scheme*
	ch2support.scm
	ch2tests.scm
	ch3support.scm
	ch4-ambeval.scm
	ch4-analyzingmceval.scm
	ch4-leval.scm
	ch4-mceval.scm
	ch4-query.scm
	ch5-compiler.scm
	ch5-eceval-compiler.scm
	ch5-eceval-support.scm
	ch5-eceval.scm
	ch5-regsim.scm
	ch5-syntax.scm
	load-eceval-compiler.scm
	load-eceval.scm

They can be loaded into the *scheme* buffer to run as a whole.

Thank you!
							Lawrence X. Amlord
							  Jul 13th, 2014
							  Xi'an, China